requirements/base.txt
pytz==2020.5  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==8.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.1.0  # https://github.com/celery/django-celery-beat
django-timezone-field==4.0  # https://github.com/celery/django-celery-beat/pull/378
flower==0.9.7  # https://github.com/mher/flower
qrbill==0.5.2
requests==2.25.1

# Django
# ------------------------------------------------------------------------------
django==3.1.7
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.10.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.12.2  # https://github.com/encode/django-rest-framework
drf-nested-routers==0.92.5
djangorestframework-datatables==0.6.0  # https://github.com/izimobil/django-rest-framework-datatables
django-cors-headers==3.6.0 # https://github.com/adamchainz/django-cors-headers
django-countries==7.0
django-ckeditor==6.0.0
django_phonenumber_field==5.0.0
phonenumbers==8.12.16
django_colorfield==0.4.1
django_localflavor==3.0.1
django-bootstrap-datepicker-plus==3.0.5
git+https://github.com/HireIQ/django-templatetag-handlebars.git@django30-support
